Change monitor command to use sketch files instead of raw filesystem manipulation
