+++ pyp5js/compiler.py
@@ -3,6 +3,7 @@ import shutil
 import shlex
 from cprint import cprint
 from unipath import Path
+from watchdog.events import PatternMatchingEventHandler
 
 
 PYP5_DIR = Path(__file__).parent
@@ -25,3 +26,22 @@ def compile_sketch_js(sketch_files):
     if sketch_files.target_dir.exists():
         shutil.rmtree(sketch_files.target_dir)
     shutil.move(__target, sketch_files.target_dir)
+
+
+class TranscryptSketchEvent(PatternMatchingEventHandler):
+    patterns = ["*.py"]
+
+    def __init__(self, *args, **kwargs):
+        self.sketch_files = kwargs.pop('sketch_files')
+        self._last_event = None
+        super().__init__(*args, **kwargs)
+
+    def on_modified(self, event):
+        event_id = id(event)
+        if event_id == self._last_event:
+            return
+
+        cprint.info(f"New change in {event.src_path}")
+        compile_sketch_js(self.sketch_files)
+        index_file = self.sketch_files.index_html
+        cprint.ok(f"Your sketch is ready and available at {index_file}")
